@@ -1,5 +1,5 @@
-name: app
-description: Ads Speed dashboard
+name: temp_package_name
+description: temp_package_description
 
 version: 1.0.0+1
 
@@ -13,22 +13,24 @@ dependencies:
     sdk: flutter
  
   equatable: ^2.0.7
-  bloc: ^8.1.4
-  flutter_bloc: ^8.1.6
-  bloc_concurrency: ^0.2.5
+  bloc: ^9.0.0
+  flutter_bloc: ^9.0.0
+  bloc_concurrency: ^0.3.0
   get_it: ^8.0.3
-  easy_localization: ^3.0.7
-  go_router: ^14.6.2
+  easy_localization: ^3.0.7+1
+  go_router: ^14.8.0
 
   cupertino_icons: ^1.0.8
-  dio: ^5.7.0
+  dio: ^5.8.0+1
   jwt_decode: ^0.3.1
-  connectivity_plus: ^6.1.1
+  connectivity_plus: ^6.1.3
   hive: ^2.2.3
   path_provider: ^2.1.5
   cached_network_image: ^3.4.1
-  flutter_svg: ^2.0.16
+  flutter_svg: ^2.0.17
   image_picker: ^1.1.2
+  image_cropper: ^9.0.0
+  flutter_keyboard_visibility: ^6.0.0
   modal_bottom_sheet: ^3.0.0
   flutter_staggered_animations: ^1.1.1
   auto_size_text: ^3.0.0
@@ -36,45 +38,28 @@ dependencies:
   file: ^7.0.1
   url_strategy: ^0.3.0
   flutter_portal: ^1.1.4
+  flutter_native_splash: ^2.4.4
   dotted_border: ^2.1.0
-  intl: ^0.18.1
   gap: ^3.0.1
   flutter_switch: ^0.3.2
-  universal_html: ^2.2.4
-  firebase_core: ^3.8.1
-  firebase_auth: ^5.3.4
-  cloud_firestore: ^5.5.1
-  cloud_functions: ^5.2.0
-  firebase_storage: ^12.3.7
-  elegant_notification: ^2.4.0 
-  desktop_drop: ^0.5.0
-  qr_flutter: ^4.1.0
-  cross_file: ^0.3.4+2
-  percent_indicator: ^4.2.4
-  flutter_image_compress: ^2.3.0
-  share_plus: ^10.1.3
-  printing: ^5.13.4
-  pdf: ^3.11.1
-  url_launcher: ^6.3.1
-  excel: ^4.0.6
-#  file_picker: ^8.1.4
-  fl_chart: ^0.69.2
-  animated_segmented_tab_control: ^2.0.0
-  horizontal_data_table: ^4.3.1
-  wrapfit: ^0.1.1
-  flutter_datetime_picker_plus: ^2.2.0
-  web3dart: ^2.7.3
+  sentry_flutter: ^8.13.0
+  provider: ^6.1.2
 
   internal_core: 
     path: internal_libs/core
+  internal_network: 
+    path: internal_libs/network
+  qr_flutter: ^4.1.0
+  firebase_core: ^3.11.0
+  firebase_auth: ^5.4.2
 
 dependency_overrides:
-  http: ^1.1.0
-  intl: ^0.18.1
-  web: ^1.0.0
+  http: ^1.3.0
+  intl: ^0.20.2
+  js: ^0.7.2
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.14.2
+  flutter_launcher_icons: ^0.14.3
   flutter_lints: ^5.0.0
 
 flutter_icons:
@@ -96,5 +81,3 @@ flutter:
     - assets/images/ 
     - assets/images/svg/ 
     - assets/images/png/ 
-    - assets/lotties/ 
-    - assets/jsons/
