@@ -1,5 +1,5 @@
 // File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
+// ignore_for_file: type=lint
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
@@ -53,11 +53,11 @@ class DefaultFirebaseOptions {
   }
 
   static const FirebaseOptions web = FirebaseOptions(
-    apiKey: "AIzaSyCo56tCNew1pp-Jcv0Hz0TiZGCKuWVrkAE",
-    authDomain: "verifyhook.firebaseapp.com",
-    projectId: "verifyhook",
-    storageBucket: "verifyhook.firebasestorage.app",
-    messagingSenderId: "824683935034",
-    appId: "1:824683935034:web:6f98775a53b0e0cf83cc03",
+    apiKey: 'AIzaSyCo56tCNew1pp-Jcv0Hz0TiZGCKuWVrkAE',
+    appId: '1:824683935034:web:6f98775a53b0e0cf83cc03',
+    messagingSenderId: '824683935034',
+    projectId: 'verifyhook',
+    authDomain: 'verifyhook.firebaseapp.com',
+    storageBucket: 'verifyhook.firebasestorage.app',
   );
 }
