@@ -1,21 +1,17 @@
-import 'package:app/src/constants/constants.dart';
 import 'package:internal_core/internal_core.dart';
 import 'package:flutter/material.dart';
 
-class WidgetAppButton extends StatelessWidget {
+class WidgetButton extends StatelessWidget {
   final bool enable;
   final bool loading;
   final String label;
-  final Widget? icon;
   final VoidCallback? onTap;
-  const WidgetAppButton({
-    super.key,
-    this.enable = true,
-    this.loading = false,
-    required this.label,
-    this.onTap,
-    this.icon,
-  });
+  const WidgetButton(
+      {super.key,
+      this.enable = true,
+      this.loading = false,
+      required this.label,
+      this.onTap});
 
   @override
   Widget build(BuildContext context) {
@@ -24,8 +20,7 @@ class WidgetAppButton extends StatelessWidget {
       child: InkWell(
         onTap: !enable || loading ? null : onTap,
         child: Ink(
-          height: 44.sw,
-          padding: EdgeInsets.symmetric(horizontal: 24.sw),
+          height: 48,
           decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(8),
               color: enable ? appColorText : appColorElement),
@@ -39,21 +34,9 @@ class WidgetAppButton extends StatelessWidget {
                       color: appColorPrimary,
                     ),
                   )
-                : Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      if (icon != null) ...[
-                        icon!,
-                        SizedBox(
-                          width: 8.sw
-                        )
-                      ],
-                      Text(
-                        label,
-                        style: w400TextStyle(
-                            color: appColorBackground, fontSize: 14.sw),
-                      )
-                    ],
+                : Text(
+                    label,
+                    style: w400TextStyle(color: appColorBackground),
                   ),
           ),
         ),
